+++ src/app/(app)/org/admin/integrations/[provider]/page.tsx
